Move bulk-import component. Move job polling to bulk-upload. remove spinner and use 1 polling spinner. remove add fees button. nav back on success.
