+++ src/app/app.routes.ts
@@ -107,7 +107,7 @@ export const routes: Routes = [
       {
         path: ':id/bulk-upload',
         loadComponent: () =>
-          import('@components/applications-list-bulk-upload/applications-list-bulk-upload.component').then(
+          import('@components/applications-list-detail/applications-list-bulk-upload/applications-list-bulk-upload.component').then(
             (m) => m.ApplicationsListBulkUpload,
           ),
       },
