--- src/app/pages/applications-list-bulk-upload/applications-list-bulk-upload.component.html
@@ -1,94 +1,0 @@
-<app-breadcrumbs
-  [items]="[
-    { label: 'Applications list', link: '/applications-list' },
-    {
-      label: 'Applications list details',
-      link: '/applications-list/' + vm().listId,
-    },
-  ]"
-/>
-
-@if (vm().fileUploadStatus === 'error') {
-  <app-error-summary
-    [title]="'There is a problem'"
-    [items]="vm().errorSummary"
-    [focusKey]="submitAttempt()"
-  />
-}
-
-@if (vm().isValidCSV === false) {
-  <app-error-summary
-    [title]="'There is a problem'"
-    [items]="[{ text: 'Please upload a valid CSV file.' }]"
-    [focusKey]="submitAttempt()"
-    targetId="file-upload-1"
-  />
-}
-
-<app-page-header title="Bulk Upload Applications" />
-
-<app-help-details summary="Help with bulk upload">
-  <p class="govuk-body">
-    Use bulk upload to add applications to this list from a CSV file.
-  </p>
-
-  <ul class="govuk-list govuk-list--bullet">
-    <li>The file must match the current bulk upload template.</li>
-    <li>The filename must end in .csv.</li>
-    <li>
-      After you upload the file, the applications list page shows the upload
-      progress.
-    </li>
-    <li>The file is validated before applications are added to the list.</li>
-    <li>
-      If the upload fails, the error message explains what needs to be
-      corrected.
-    </li>
-  </ul>
-</app-help-details>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <p class="govuk-body">
-      Select the bulk applications file you wish to upload.
-    </p>
-
-    @if (!vm().isUploadInProgress) {
-      <div class="govuk-form-group remove-bot-margin govuk-width-container">
-        <form (ngSubmit)="onSubmit()">
-          <label class="govuk-label" for="file-upload-1"> Upload a file </label>
-          <div data-module="govuk-file-upload">
-            <input
-              accept=".csv,text/csv"
-              (change)="onFileSelected($event)"
-              class="govuk-file-upload"
-              id="file-upload-1"
-              name="fileUpload1"
-              type="file"
-            />
-            <button
-              type="submit"
-              class="govuk-button govuk-!-margin-top-5"
-              [disabled]="!vm().isValidCSV"
-              data-module="govuk-button"
-            >
-              Upload file
-            </button>
-          </div>
-        </form>
-      </div>
-    } @else {
-      <app-loading-spinner [text]="'Uploading file, please wait...'" />
-    }
-  </div>
-</div>
-<!-- Comment out for now -->
-<!-- <div class="govuk-form-group">
-  <button
-    class="govuk-button float-right"
-    [disabled]="true"
-    data-module="govuk-button"
-  >
-    Add Fee Details
-  </button>
-</div> -->
