+++ src/app/pages/applications-list-detail/applications-list-bulk-upload/applications-list-bulk-upload.component.html
@@ -1,0 +1,112 @@
+<app-breadcrumbs
+  [items]="[
+    { label: 'Applications list', link: '/applications-list' },
+    {
+      label: 'Applications list details',
+      link: '/applications-list/' + vm().listId,
+    },
+  ]"
+/>
+
+@if (vm().fileUploadStatus === 'error') {
+  <app-error-summary
+    [title]="'There is a problem'"
+    [items]="vm().errorSummary"
+    [focusKey]="submitAttempt()"
+  />
+}
+
+@if (vm().isValidCSV === false) {
+  <app-error-summary
+    [title]="'There is a problem'"
+    [items]="[{ text: 'Please upload a valid CSV file.' }]"
+    [focusKey]="submitAttempt()"
+    targetId="file-upload-1"
+  />
+}
+
+@if (vm().bulkUploadFeedback?.kind === 'progress' || vm().isUploadInProgress) {
+  <app-async-job-progress
+    [heading]="vm().bulkUploadFeedback?.heading ?? 'Upload in progress'"
+    [body]="vm().bulkUploadFeedback?.body ?? 'Uploading file, please wait...'"
+  />
+}
+<!-- move to app-list-detail -->
+<!-- @if (vm().bulkUploadFeedback?.kind === 'success') {
+  <app-success-banner
+    [heading]="vm().bulkUploadFeedback?.heading ?? ''"
+    [body]="vm().bulkUploadFeedback?.body ?? ''"
+  />
+} -->
+
+@if (vm().bulkUploadFeedback?.kind === 'warning') {
+  <app-notification-banner
+    variant="warning"
+    [title]="vm().bulkUploadFeedback?.title ?? 'Warning'"
+    [heading]="vm().bulkUploadFeedback?.heading ?? ''"
+    [body]="vm().bulkUploadFeedback?.body ?? ''"
+    [autoFocus]="true"
+  />
+}
+
+@if (vm().bulkUploadFeedback?.kind === 'error') {
+  <app-error-summary
+    [title]="vm().bulkUploadFeedback?.heading ?? 'There is a problem'"
+    [items]="[{ text: vm().bulkUploadFeedback?.body ?? '' }]"
+    [focusKey]="submitAttempt()"
+  />
+}
+
+<app-page-header title="Bulk Upload Applications" />
+
+<app-help-details summary="Help with bulk upload">
+  <p class="govuk-body">
+    Use bulk upload to add applications to this list from a CSV file.
+  </p>
+
+  <ul class="govuk-list govuk-list--bullet">
+    <li>The file must match the current bulk upload template.</li>
+    <li>The filename must end in .csv.</li>
+    <li>
+      After you upload the file, the applications list page shows the upload
+      progress.
+    </li>
+    <li>The file is validated before applications are added to the list.</li>
+    <li>
+      If the upload fails, the error message explains what needs to be
+      corrected.
+    </li>
+  </ul>
+</app-help-details>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <p class="govuk-body">
+      Select the bulk applications file you wish to upload.
+    </p>
+
+    <div class="govuk-form-group remove-bot-margin govuk-width-container">
+      <form (ngSubmit)="onSubmit()">
+        <label class="govuk-label" for="file-upload-1"> Upload a file </label>
+        <div data-module="govuk-file-upload">
+          <input
+            accept=".csv,text/csv"
+            (change)="onFileSelected($event)"
+            class="govuk-file-upload"
+            id="file-upload-1"
+            name="fileUpload1"
+            type="file"
+          />
+        </div>
+        <button
+          type="submit"
+          class="govuk-button govuk-!-margin-top-5"
+          [disabled]="!vm().isValidCSV || vm().isUploadInProgress"
+          data-module="govuk-button"
+        >
+          Upload file
+        </button>
+      </form>
+    </div>
+  </div>
+</div>
