+++ src/app/pages/applications-list-detail/applications-list-bulk-upload/applications-list-bulk-upload.component.scss
