+++ src/app/pages/applications-list-detail/applications-list-detail.component.html
@@ -54,38 +54,6 @@
   />
 }
 
-@if (vm().bulkUploadFeedback?.kind === 'progress') {
-  <app-async-job-progress
-    [heading]="vm().bulkUploadFeedback?.heading ?? ''"
-    [body]="vm().bulkUploadFeedback?.body ?? ''"
-  />
-}
-
-@if (vm().bulkUploadFeedback?.kind === 'success') {
-  <app-success-banner
-    [heading]="vm().bulkUploadFeedback?.heading ?? ''"
-    [body]="vm().bulkUploadFeedback?.body ?? ''"
-  />
-}
-
-@if (vm().bulkUploadFeedback?.kind === 'warning') {
-  <app-notification-banner
-    variant="warning"
-    [title]="vm().bulkUploadFeedback?.title ?? 'Warning'"
-    [heading]="vm().bulkUploadFeedback?.heading ?? ''"
-    [body]="vm().bulkUploadFeedback?.body ?? ''"
-    [autoFocus]="true"
-  />
-}
-
-@if (vm().bulkUploadFeedback?.kind === 'error') {
-  <app-error-summary
-    [title]="vm().bulkUploadFeedback?.heading ?? 'There is a problem'"
-    [items]="[{ text: vm().bulkUploadFeedback?.body ?? '' }]"
-    [focusKey]="submitAttempt()"
-  />
-}
-
 <app-page-header
   title="Applications list"
   [actions]="[
